requirements.txt
pybit>=5.8.0
psycopg2-binary>=2.9.0
